cadre-model-backend/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
marshmallow==3.20.1
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
openpyxl==3.1.2
pymysql==1.1.0
cryptography==41.0.7

